Remove legacy frontend build path
