+++ frontend/react/app.ts
@@ -3,11 +3,17 @@
 
 // Import theme toggle (runs on all pages)
 import './theme-toggle.ts';
+import './components/user-menu.ts';
 
 // Import page-specific components
 // These check for their mount points and only initialize if found
 import './pages/gallery.tsx';
 import './pages/image-detail.tsx';
+import './pages/gallery-preview-template.ts';
+import './pages/login.ts';
+import './pages/passkey-enrollment.ts';
+import './pages/passkeys.ts';
+import './pages/profile.ts';
 
 // Export a global object for any runtime configuration if needed
 (window as any).Tenrankai = {
