+++ templates/pages/passkeys.html.liquid
@@ -36,102 +36,4 @@
     </div>
 </div>
 
-<!-- Load TypeScript modules for comprehensive passkey management -->
-<script type="module">
-  import { WebAuthnUtils, PasskeyManager } from '{{ "js/components/auth/webauthn.js" | asset_url }}';
-  import { AuthUtils } from '{{ "js/components/auth/auth-utils.js" | asset_url }}';
-  import { Toast } from '{{ "js/components/ui/toast.js" | asset_url }}';
-  
-  // Load existing passkeys
-  async function refreshPasskeysList() {
-      const listContainer = document.getElementById('passkeysList');
-      
-      try {
-          const passkeys = await PasskeyManager.loadPasskeys();
-          
-          if (passkeys.length === 0) {
-              listContainer.innerHTML = '<p class="no-passkeys">No passkeys registered yet.</p>';
-          } else {
-              listContainer.innerHTML = passkeys.map(pk => `
-                  <div class="passkey-item" data-id="${pk.id}">
-                      <div class="passkey-info">
-                          <div class="passkey-name">${AuthUtils.escapeHtml(pk.name)}</div>
-                          <div class="passkey-meta">
-                              Created: ${PasskeyManager.formatDate(pk.created_at)}
-                              ${pk.last_used_at ? `• Last used: ${PasskeyManager.formatDate(pk.last_used_at)}` : ''}
-                          </div>
-                      </div>
-                      <button class="btn-delete" data-passkey-id="${pk.id}">Delete</button>
-                  </div>
-              `).join('');
-              
-              // Add event listeners for delete buttons
-              const deleteButtons = listContainer.querySelectorAll('[data-passkey-id]');
-              deleteButtons.forEach(button => {
-                  const passkeyId = button.getAttribute('data-passkey-id');
-                  if (passkeyId) {
-                      button.addEventListener('click', () => handleDeletePasskey(passkeyId));
-                  }
-              });
-          }
-      } catch (error) {
-          listContainer.innerHTML = '<p class="error">Failed to load passkeys</p>';
-      }
-  }
-
-  // Delete passkey handler
-  async function handleDeletePasskey(passkeyId) {
-      if (!confirm('Are you sure you want to delete this passkey?')) {
-          return;
-      }
-      
-      try {
-          await PasskeyManager.deletePasskey(passkeyId);
-          Toast.success('Passkey deleted successfully!');
-          refreshPasskeysList(); // Reload the list
-      } catch (error) {
-          Toast.error('Failed to delete passkey. Please try again.');
-      }
-  }
-
-  // Add new passkey
-  document.getElementById('addPasskeyForm').addEventListener('submit', async function(e) {
-      e.preventDefault();
-      
-      const nameInput = document.getElementById('passkeyName');
-      const name = nameInput.value.trim();
-      const submitButton = e.target.querySelector('button[type="submit"]');
-      
-      if (!name) {
-          Toast.error('Please enter a name for the passkey');
-          return;
-      }
-      
-      submitButton.disabled = true;
-      submitButton.textContent = 'Adding...';
-      
-      try {
-          await WebAuthnUtils.registerPasskey(name);
-          Toast.success('Passkey added successfully!');
-          nameInput.value = '';
-          refreshPasskeysList(); // Reload the list
-          
-      } catch (error) {
-          Toast.error(error.message || 'Failed to add passkey. Please try again.');
-      } finally {
-          submitButton.disabled = false;
-          submitButton.textContent = 'Add Passkey';
-      }
-  });
-
-  // Check WebAuthn support
-  if (!WebAuthnUtils.isSupported()) {
-      document.querySelector('.add-passkey-section').innerHTML = 
-          '<div class="message info">Your browser does not support passkeys.</div>';
-  }
-
-  // Load passkeys on page load
-  refreshPasskeysList();
-</script>
-
 {% include "_footer.html.liquid" %}
