--- tsconfig.legacy.json
@@ -1,31 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
-    "module": "Node16",
-    "moduleResolution": "node16",
-    "outDir": "./static/js",
-    "rootDir": "./frontend/legacy",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "allowJs": false,
-    "declaration": false,
-    "removeComments": true,
-    "sourceMap": true,
-    "paths": {
-      "@/*": ["./frontend/legacy/*"]
-    }
-  },
-  "include": [
-    "frontend/legacy/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "static",
-    "target",
-    "dist"
-  ]
-}
